Incomplete workaround of stream task hanging
Wakeup counter can "wrap" back to sleep.  Need to understand exactly why
we're posting that many events.  Any workaround needs to be safe for
cases when the excessive events are posting when we're not yet slept
(e.g. task counter getting "stuck" when we're running is bad, because
then we can block on a higher value and miss the events).